gltf: Move material functionality into material.cpp
Also move usesTextureSet into material.cpp since that's where
material-specific logic actually is necessary.
